Added code generation for windows.
Added a generationWin.hpp that can generate valid assembly for windows. I have not added a windows linker into the project because that would mean adding an exe to the project which could make it look unsafe. So you can choose your own linker. I prefer GoLink as that's what I used to test this.